commit update

+++ src/main/java/com/foodDelivery/project/service/impl/ReviewServiceImpl.java
@@ -1,13 +1,13 @@
 package com.foodDelivery.project.service.impl;
 
 import com.foodDelivery.project.domen.dto.ReviewDTO;
-import com.foodDelivery.project.domen.dto.UserDTO;
 import com.foodDelivery.project.domen.model.Order;
-import com.foodDelivery.project.domen.model.Product;
 import com.foodDelivery.project.domen.model.Review;
+import com.foodDelivery.project.domen.model.User;
 import com.foodDelivery.project.domen.responce.ReviewToRetrieve;
 import com.foodDelivery.project.exception.BusinessException;
 import com.foodDelivery.project.repository.ReviewRepository;
+import com.foodDelivery.project.repository.UserRepository;
 import com.foodDelivery.project.service.ReviewService;
 import jakarta.transaction.Transactional;
 import org.slf4j.Logger;
@@ -17,36 +17,45 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
 //все методы изменения review сделать по пользователю
-//понять preAuthorize
+//поменять preAuthorize
 @Service
 @PreAuthorize(value = "hasRole('ROLE_ADMIN') or hasRole('ROLE_USER')")
 public class ReviewServiceImpl implements ReviewService {
 
     private ReviewRepository repository;
 
+    private final UserRepository userRepository;
+
     private static final Logger log = LoggerFactory.getLogger(ReviewServiceImpl.class);
     @Autowired
-    public ReviewServiceImpl(ReviewRepository repository) {
+    public ReviewServiceImpl(ReviewRepository repository, UserRepository userRepository) {
         this.repository = repository;
+        this.userRepository = userRepository;
     }
 
-    //удалить
-    @Override
-    public void createReview(ReviewDTO reviewDTO){
-        Review review = new Review();
-        review.setRating(reviewDTO.getRating());
-        review.setComment(reviewDTO.getComment());
-        repository.save(review);
-        log.info("Отзыв успешно добавлен.");
+    private User getCurrentUser(){
+        Authentication authentication =
+                SecurityContextHolder
+                        .getContext()
+                        .getAuthentication();
+        String username = authentication.getName();
+        return userRepository.findUserByUsername(username)
+                .orElseThrow(() -> new BusinessException(
+                        "Пользователь не найден!",
+                        HttpStatus.NOT_FOUND
+                ));
     }
 
     @Override
     @Transactional
+    @PreAuthorize("hasRole('ROLE_USER')")
     public Review createReviewWithOrder(ReviewDTO reviewDTO, Order order) {
         Review review = new Review();
 
@@ -55,25 +64,40 @@ public class ReviewServiceImpl implements ReviewService {
         review.setOrder_id(order);
         review.setUser_id(order.getUser_id());
 
-        return repository.save(review);
+        Review saved = repository.save(review);
+        log.info("Отзыв успешно создан вместе с заказом");
+
+        return saved;
     }
 
     @Override
+    @PreAuthorize("isAuthenticated()")
     public List<ReviewToRetrieve> getReviews(){
         List<Review> all = repository.findAll();
+
+        if(all.isEmpty()){
+            throw new BusinessException(
+                    "Отзывы не найдены",
+                    HttpStatus.NOT_FOUND );
+        }
 
         List<ReviewToRetrieve> reviewToRetrieves = new ArrayList<>();
 
         for(Review review : all){
             ReviewToRetrieve reviewToRetrieve = new ReviewToRetrieve();
+
             reviewToRetrieve.setComment(review.getComment());
             reviewToRetrieve.setRating(review.getRating());
+
             reviewToRetrieves.add(reviewToRetrieve);
         }
+
+        log.info("Список отзывов успешно получен");
         return reviewToRetrieves;
     }
 
     @Override
+    @PreAuthorize("isAuthenticated()")
     public ReviewToRetrieve getReviewById(Long id) {
         Review review = repository.findById(id)
                 .orElseThrow(() -> new BusinessException(
@@ -82,6 +106,7 @@ public class ReviewServiceImpl implements ReviewService {
                 ));
 
         ReviewToRetrieve dto = new ReviewToRetrieve();
+
         dto.setComment(review.getComment());
         dto.setRating(review.getRating());
 
@@ -90,6 +115,7 @@ public class ReviewServiceImpl implements ReviewService {
     }
 
     @Override
+    @PreAuthorize("isAuthenticated()")
     public List<ReviewToRetrieve> findReviewsWithPageble(PageRequest of) {
         Page<Review> page = repository.findAll(of);
 
@@ -97,8 +123,10 @@ public class ReviewServiceImpl implements ReviewService {
 
         for (Review review : page.getContent()) {
             ReviewToRetrieve dto = new ReviewToRetrieve();
+
             dto.setComment(review.getComment());
             dto.setRating(review.getRating());
+
             result.add(dto);
         }
 
@@ -107,12 +135,21 @@ public class ReviewServiceImpl implements ReviewService {
     }
 
     @Override
+    @PreAuthorize("hasRole('ROLE_USER')")
     public ReviewDTO updateReview(Long id, ReviewDTO reviewDTO) {
         Review review = repository.findById(id)
                 .orElseThrow(() -> new BusinessException(
                         "Отзыв не найден",
                         HttpStatus.NOT_FOUND
                 ));
+
+        User currentUser = getCurrentUser();
+
+        if (!review.getUser_id().getId().equals(currentUser.getId())) {
+            throw new BusinessException(
+                    "Это не ваш отзыв",
+                    HttpStatus.FORBIDDEN );
+        }
 
         review.setComment(reviewDTO.getComment());
         review.setRating(reviewDTO.getRating());
@@ -121,18 +158,31 @@ public class ReviewServiceImpl implements ReviewService {
         Review saved = repository.save(review);
 
         ReviewDTO dto = new ReviewDTO();
+
         dto.setComment(saved.getComment());
         dto.setRating(saved.getRating());
         dto.setCreatedAt(saved.getCreatedAt());
 
+        log.info("Отзыв с id {} успешно обновлен", id);
         return dto;
     }
 
     @Override
+    @PreAuthorize("hasRole('ROLE_USER')")
     public void delete(Long id) {
         Review review = repository.findById(id)
                 .orElseThrow(() -> new BusinessException("Отзыв не найден!", HttpStatus.NOT_FOUND));
+
+        User currentUser = getCurrentUser();
+
+        if(!review.getUser_id().getId().equals(currentUser.getId())){
+            throw new BusinessException(
+                    "Это не ваш отзыв!",
+                    HttpStatus.FORBIDDEN
+            );
+        }
         repository.delete(review);
+
         log.info("Отзыв с id {} успешно удален.", id);
     }
 }
